Rename metrics benchmarks
